Fix comment watching
Since we moved to NSFilePresenters we were only watching for changes to comment files, not
for additions. We now watch properly for files added. This needs a bit of cleaning up in
terms of the duplication around user comment loading. We also don't handle the move-replace
cycle for non-user comments (as the existing file presenter sticks to the old URL until
it is properly deleted). Will need to see if this causes any practical problems with
sync'd folders.