Work around presentedSubitemDidAppear never being called
For reasons unknown, presentedSubitemDidAppear is never called. This means that we weren't
picking up new audio files in any folders. Instead, we watch for presentedSubitemDidChange.
This gets called for atime changes too, so we need to be more pro-active in filtering those
out.

Maybe, just maybe, this move to Coordinated access will actually have all the holes covered
now and well be back to where we were months ago!