Ignore sidecar dir updates in audio watchers
This coordinated access thing is driving me nuts. If access outside the app is
not coordinated, we don't get notifications, so we have to be over zealous. But
then, we get proper notifications if it is, so being over zealous bites us.
Many apps (eg: logic) don't use coordinated access for bounce to disk etc.
Hopefuly this fixes #3 (again) without any other bonus side effects.